runnero/requirements.txt
django >= 3.1
psycopg2-binary == 2.8.5
djangorestframework >= 3.12.1
djangorestframework_simplejwt